Energy ball zaps fixed & inline with modern power.
This is the first step to the return of TG's lost engines.  Teslas are simple to setup, but can't scale like a Supermatter can short of making more of them - and unlike a SM with its large damage buffer, all you need to have a containment breach with a tesla engine is field failure, making them decidedly riskier.