Scythes now fit in coroner labcoat/wintercoat suit storage (#77960)
## About The Pull Request
Allows the coroner to store the scythe in the suit storage of their
labcoat and/or wintercoat.

![image](https://github.com/tgstation/tgstation/assets/44104681/725caf4c-fe45-439c-bc61-7c33a4fd3218)

(skeleton model dressed just for fun, not relevant to changes)
## Why It's Good For The Game
Enables more flavorful coroner gameplay without giving up your backpack
for a very mediocre weapon. The scythe does the same damage as the
circular saw or ritual knife, has less armor pen than some of the
nullrods, and less wounding potential than the circular saw or ritual
knife. It is a lethal weapon, but it hardly wins any awards, and isn't
quite worth giving up your bag for. I think the atmosphere surrounding
coroner benefits from the presence of the scythe, and the potential it
has as a weapon simply does not warrant how discouraged you are from
having it on your person.
## Changelog
:cl:
qol: The coroner can now store scythes in their labcoat and winter coat
suit storage. Ask not for whom the bell tolls...
/:cl: