V0.1 of PrairieOutpostMkII, Inspired by Rimpoint
A complete rework to make PrairieOutpost into a dual-Z map, with an upper area built as an open-layout "town" with the various departments as standalone buildings, connected by a roadway bearing the atmos and power lines - and a lower area meant to be the primary mining level, with greater threats, no natural light, and minimal infrastructure - but greater rewards should you dare to delve.  This is an early version, featuring only Atmospherics, Arrivals, Departures, and an early Cargo - not ready for play yet.

This also introduces a small mapping variable helper for atmos machines, allowing you to *forcibly override* their init directions to emulate the functionality from an RPD; smart pipes are actually pretty dumb when it comes to how they play when mapping, and it annoyed me too much to leave alone.