GMM UI & Cargo budget handling tweaks (#79145)
## About The Pull Request
- Fixes #79115

This PR is a 2-part solution as specified by the last paragraph of the
above-mentioned issue.

> but also, for you to be prevented from adding more to an order that
you don't reasonably have the funds in that account to afford.

The GMM will now not add stacks to an existing order if it exceeds the
available (private or cargo depending on the mode of ordering) budget or
if it exceeds the available market quantity on the market.

The UI is improved.
- Much larger to display all materials without scrolling
- Now shows the total cost of the order
- Display's number of sheets ordered next to each material type
- Has a clear button to cancel the order
- Order buttons greys out if the order quantity is greater than what's
available in the market or if you don't have the available budget for it
![Screenshot
(340)](https://github.com/tgstation/tgstation/assets/110812394/b2e0dc50-9eee-457f-a85e-c011de014cc6)

But this is not enough. The below example will explain why.

Say you have a budget of 500 cr in both private & cargo.
1. Place an order of materials worth 500 cr be it private or by cargo
budget.
2. Now go spend that budget on something such that it falls below 500
cr, say in this example you overspend, and the available budget becomes
0(you broke)
3. Now go call the shuttle. The order gets rejected but it's left
hanging in the cargo checkout queue and will only be cleared till you
raise your budget above 500 cr and again call the shuttle.

This example just applies to 500 cr but with larger amounts you order
will be hanging in the queue indefinitely till your broke... makes the
available budget to finally purchase it

To fix this we go to the 2nd part of the paragraph.

> An ideal fix would be for GMM orders to be cancellable.

cargo will now **"remove/cancel"** this order and will not leave it
hanging in the queue indefinitely. So, this way if you by accident made
a material order way beyond your available budget cargo will forgive you
and clear you of that debt so you can start a fresh order.

Of course, this fix is only useful if you send the order without
checking the budget but even if you do there is plan B so it's helpful
to know.

## Changelog
:cl:
fix: Cargo will remove/cancel orders from its cart if that order exceeds
the available budget (both private or cargo) and the player cannot
cancel this order manually. All order costs are rounded up to integer
values
fix: Galactic material market will deny appending stacks to your
existing order if it exceeds the available (private or cargo depending
on the mode of ordering) budget & if it exceeds the available materials
on the market. Galactic material market UI is overall improved.
/:cl:

---------

Co-authored-by: ArcaneMusic <41715314+ArcaneMusic@users.noreply.github.com>
Co-authored-by: Jacquerel <hnevard@gmail.com>