adds screentips to the janicart (pimpin' ride) (#80665)
## About The Pull Request
Adds screentips the janicart (pimpin' ride) for dismounting, adding and
removing trash bags/upgrades/keys and inserting into the connected trash
bag.

The dismounting and key insertion/removal screentips could maybe be used
in the base vehicle. But right now I can't figure out how to move those
there without needing to do the same checks on the janicart (pimpin'
ride) again.
## Why It's Good For The Game
The janicart (pimpin' ride) has a lot of things you can do with it that
are not immediately obvious. It took me a while to figure out that you
can grab the trash bag from it without having to get off of it first.
## Changelog
:cl:
qol: added screentips to the janicart (pimpin' ride)
/:cl: