Respriting the goldfish and lanternfish (#81148)
## About The Pull Request
Those are two fishes I was too tired to resprite as well last time. I've
got to say their color schemes came out a bit more lively than the rest.

## Why It's Good For The Game

![immagine](https://github.com/tgstation/tgstation/assets/42542238/971d1171-317a-4ebc-903e-c9378c2065a2)

The pufferfish looks too cute as is so I couldn't bring myself to
resprite it.

## Changelog

:cl:
image: Resprited the goldfish and lanternfish as well.
/:cl: