Adds a 4th law to the Painter lawset regarding cultivating an audience (#81717)
## About The Pull Request
This PR adds the following law to the Artist lawset:

`4. Art requires appreciation. Cultivate an audience aboard the station
to ensure as many as possible see your works.`

## Why It's Good For The Game
Artist has always been kind of vague (intentionally, I imagine), but I
don't think the intent has ever been to encourage AIs to go "Blood is
art, right? It constitutes art if I just murder fucking everyone on the
station, right?". This law helps to steer things in a more constructive
direction while leaving room for loopholes. I tried to word it in a way
that allows AIs who want to be a little daring and bloodthirsty to kill
a crewmember or two for "paint" without encouraging them to kill
everyone. You need an audience, so killing everyone on the station would
run counter to that.

## Changelog
:cl:
add: Added a new law to the Artist lawset in order to encourage Artist
AIs to build an audience.
/:cl: