[MIRROR] Adjusts some of the contraband weights for Contraband Crates. (#2846)
* Adjusts some of the contraband weights for Contraband Crates. (#81771)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This makes getting Contraband Plus (the better stuff with more
interesting effects), slightly higher in probability out of contraband
crates.

It also removes the completely redundant narcotics roll, and the truly
awful permabrig contraband roll, from Contraband Plus.

Some of the more interesting narcotics are now directly in Contraband
Plus.

## Why It's Good For The Game

The weights were a little uninteresting. You rarely if ever saw the
effects of the contraband crate change with the current values. It was
kind of lame to see a shiv in a contraband crate knowing that it was a
wasted Contraband Plus roll. This hopefully stops that from happening
too often.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Adjusts the values and contents of the Contraband Crate's item
loot table.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: san7890 <the@san7890.com>

* (upstream PR 81771)

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>
Co-authored-by: san7890 <the@san7890.com>