[MIRROR] Adds a missing recharger to the Birdshot science testing range (#2895)
* Adds a missing recharger to the Birdshot science testing range (#81957)

## About The Pull Request
As per the title, adds a recharger to the testing range in science on
Birdshot.
I noticed that this was missing from the room, and without it there's no
way to recharge the nearby practice weaponry. So I added one for the
practice weapons and for the consistency with testing ranges in other
maps.
## Why It's Good For The Game
If you have a firing range with (practice) weaponry and no way to
recharge them after firing, it kind of sucks. By adding one, we both fix
this problem and the inconsistency with other testing ranges.
## Changelog
:cl:
fix: A missing recharger has been added into science's testing range on
Birdshot station.
/:cl:

* (upstream PR 81957)

---------

Co-authored-by: Derpguy3 <85206687+Derpguy3@users.noreply.github.com>